requirements.txt
# LiveKit Agents with OpenAI, Google (Gemini), and Silero VAD plugins
# Python 3.10+ required (3.9 is not supported)
livekit-agents[openai,google,silero]>=1.3.0
python-dotenv>=1.0.0

# Token server for frontend (self-hosted)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
